Move ImageType enum to bci_tester.data
